rename package form metal_sdk to metal
